==> 01_scraping.ipynb <==
# Project 3 - Creation vs Evolution

## Problem Statement

In this project, I use Reddit's API to scrape data from two subreddits, r/Evolution and r/Creation.  My goal is to build a model which may accurately and effectively predict and categorize post titles by subreddit.  Using NLP techniques, how accurately can we predict the subreddit based on the post titles?

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

from IPython.core.display import HTML
import pylast
import requests
import time

### Importing json data from _reddit.com_

creation_url = 'https://www.reddit.com/r/Creation.json'
evolution_url = 'https://www.reddit.com/r/evolution.json'

headers = {'User-agent': 'harmon get 0.1'}

res_c = requests.get(creation_url, headers=headers)
res_c.status_code

res_e = requests.get(evolution_url, headers=headers)
res_e.status_code

creation_json = res_c.json()

evolution_json = res_e.json()

### Extracting from the JSON

sorted(creation_json.keys())

sorted(creation_json['data'].keys())

# returning list of 'names' of posts
[post['data']['name'] for post in creation_json['data']['children']]

pd.DataFrame(creation_json['data']['children'])

creation_json['data']['children'][0]['data'].keys()

creation_json['data']['after']

### Getting multiple pages of posts from r/creation
    - Scraping performed on 3/31/2019

creation_posts = []
after = None
for i in range(50):     # 50 pages of posts
    if i % 5 == 0:
        print(f'Parsing page number {i}')
    if after == None:
        params = {}
    else:
        params = {'after': after}
    creation_url = 'https://www.reddit.com/r/Creation.json'
    res_c = requests.get(creation_url, params=params, headers=headers)
    if res_c.status_code == 200:
        creation_json = res_c.json()
        creation_posts.extend(creation_json['data']['children'])
        after = creation_json['data']['after']
    else:
        print(res_c.status_code)
        print('Error fetching reddit json')
        break
    time.sleep(3)       # don't want to be labeled as a DDOS attack

len(creation_posts)

# are there any duplicate posts?
len(set([post['data']['title'] for post in creation_posts]))

# dataframe from set of unique post titles
crea_df = pd.DataFrame(set([post['data']['title'] for post in creation_posts]))

crea_df.head(3)

crea_df.rename(columns={0:'title'}, inplace=True)

crea_df['is_Evolution'] = 0

crea_df.head(3)

### Getting multiple pages of posts from r/evolution
    - Scraping performed on 3/31/2019

evolution_posts = []
after = None
for i in range(50):
    if i % 5 == 0:
        print(f'Parsing page number {i}')
    if after == None:
        params = {}
    else:
        params = {'after': after}
    evolution_url = 'https://www.reddit.com/r/evolution.json'
    res_e = requests.get(evolution_url, params=params, headers=headers)
    if res_e.status_code == 200:
        evolution_json = res_e.json()
        evolution_posts.extend(evolution_json['data']['children'])
        after = evolution_json['data']['after']
    else:
        print(res_e.status_code)
        print('Error fetching reddit json')
        break
    time.sleep(3)

len(evolution_posts)

# are there any duplicate posts?
len(set([post['data']['title'] for post in evolution_posts]))

# dataframe from set of unique post titles
evo_df = pd.DataFrame(set([post['data']['title'] for post in evolution_posts]))

evo_df.head(3)

evo_df.rename(columns={0:'title'}, inplace=True)

evo_df['is_Evolution'] = 1

evo_df.head(3)

### Combining post title dataframes

df = evo_df.append(crea_df)

df.head(3)

df.tail(3)

The dataframe should have 1969 rows, but the tail shows an index only to 994.  This issue resulted from combining the two dataframes with `.append()`, and currently most of our indexes apply to two separate rows.

To resolve this, we use `.reset_index(drop=True, inplace=True)` to reset the index for our new, combined dataframe!

# reseting the indexes to have fully unique indexes
df.reset_index(drop=True, inplace=True)

df.head(3)

df.tail(3)

### Saving the "raw" dataframe

df.to_csv('./data/raw_titles.csv')
