==> 02_cleaning_tokenizing.ipynb <==
# Part 2 - Cleaning/Tokenizing

In this portion of the report, I take the raw post titles from the JSONs and clean/tokenize them.  This will allow vectoriztion in the next notbook for NLP.  Additionally, I chose to complete this in a separate notebook because it will allow me to return and apply stemming or lemmatization, should I choose to apply that to my models.

import pandas as pd

df_raw = pd.read_csv('./data/raw_titles.csv')

df_raw.head()

df_raw.drop(columns='Unnamed: 0', inplace=True)

df_raw.head()

## Time to Transform

I will be tokenizing each of the titles in order to analyze the most indicative words, but I am chosing to **_NOT_** stem or lemmatize the tokens because I believe that tense may be a powerful indicator for these two subreddits, given their context.

If I find that tense is not as powerful an indicator as I believed, I will plan to perform a second analysis with the stemmed/lemmatized tokens.

import re
from nltk.tokenize import RegexpTokenizer
from nltk.corpus import stopwords

tokenizer = RegexpTokenizer(r'\w+')
tokenizer.tokenize(df_raw['title'][0].lower())

def title_to_words(raw_review):
    tokenizer = RegexpTokenizer(r'\w+')
    words = tokenizer.tokenize(title.lower())    # tokenizing the title
    stops = set(stopwords.words('english'))                   # setting stop words
    meaningful_words = [w for w in words if not w in stops]   # removing stop words from titles
    return(" ".join(meaningful_words))                        # return the title as a list

clean_titles = []
for title in df_raw['title']:
    clean_titles.append(title_to_words(title))

df_clean = pd.DataFrame(clean_titles)

df_clean.rename(columns={0:'title'}, inplace=True)

df_clean['is_evolution'] = df_raw['is_Evolution']

df_clean.head()

df_clean.loc[1566]

The removal of my stop words removed every word from the post at index 1566, resulting in an empty string.  This caused a null value to be introduced when I imported the data into a new notebook.  I have removed the row in order to resolve this.

df_clean.drop(index=1566, inplace=True)

df_clean.to_csv('./data/clean_token_titles.csv', index=False)

In the next notebook, I will be vectorizing the data and performing EDA!
